Create Logic basic logic and table for the game
In this commit the app was add a grid and a input for the user,
now the user can enter the word and the result will be generated and
displayed by the grid component.
States added in App.js: grid, guess, word

Note: This commit is just a prototype and not the main logic the logic
will be seperated to other files in upcomming commits.